Do not import networkx (and transitively numpy) until it is needed
This speeds up `leanproject --version` from 3.4s to 1.9s on my machine
